refactor: close ui architecture review findings
