+++ src/niamoto/gui/ui/src/app/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import '@/index.css'
-import '@/shared/i18n'
+import '@/i18n'
 import App from './App'
 import { RootProviders } from './providers/RootProviders'
