+++ src/niamoto/gui/ui/src/features/site/components/FooterSectionsEditor.tsx
@@ -38,7 +38,7 @@ import type {
   FooterLink,
   StaticPage,
   GroupInfo,
-} from '@/features/site/hooks/useSiteConfig'
+} from '@/shared/hooks/useSiteConfig'
 
 interface FooterSectionsEditorProps {
   sections: FooterSection[]
