+++ src/niamoto/gui/ui/src/features/site/components/GroupPageViewer.tsx
@@ -14,7 +14,7 @@ import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/com
 import { Button } from '@/components/ui/button'
 import { Badge } from '@/components/ui/badge'
 import { Separator } from '@/components/ui/separator'
-import type { GroupInfo } from '@/features/site/hooks/useSiteConfig'
+import type { GroupInfo } from '@/shared/hooks/useSiteConfig'
 
 interface GroupPageViewerProps {
   group: GroupInfo
