+++ src/niamoto/gui/ui/src/features/site/components/MultilingualMarkdownEditor.tsx
@@ -28,7 +28,7 @@ import { Button } from '@/components/ui/button'
 import { Badge } from '@/components/ui/badge'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { MarkdownEditor } from './MarkdownEditor'
-import { useUpdateFileContent } from '@/features/site/hooks/useSiteConfig'
+import { useUpdateFileContent } from '@/shared/hooks/useSiteConfig'
 import { useLanguages } from '@/shared/contexts/LanguageContext'
 
 interface LanguageFileState {
