+++ src/niamoto/gui/ui/src/features/site/components/forms/GlossaryForm.tsx
@@ -21,7 +21,7 @@ import { LocalizedInput, type LocalizedString } from '@/components/ui/localized-
 import {
   useDataContent,
   useUpdateDataContent,
-} from '@/features/site/hooks/useSiteConfig'
+} from '@/shared/hooks/useSiteConfig'
 
 // Types for glossary.html context
 interface TermItem {
