+++ src/niamoto/gui/ui/src/features/site/components/forms/ImagePickerField.tsx
@@ -1,11 +1,11 @@
 /**
- * FilePickerField - File selector with upload capability
+ * ImagePickerField - Image selector with upload capability and preview
  *
- * Allows selecting existing files from a folder or uploading new ones
+ * Allows selecting existing images from a folder or uploading new ones
  */
 
 import { useRef, useState } from 'react'
-import { Upload, FolderOpen, Link, Loader2, File, X } from 'lucide-react'
+import { Upload, FolderOpen, Link, Loader2, Image as ImageIcon, X } from 'lucide-react'
 import { toast } from 'sonner'
 import { Input } from '@/components/ui/input'
 import { Button } from '@/components/ui/button'
@@ -17,38 +17,26 @@ import {
 import { ScrollArea } from '@/components/ui/scroll-area'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { cn } from '@/lib/utils'
-import { useProjectFiles, useUploadFile } from '@/features/site/hooks/useSiteConfig'
+import { useProjectFiles, useUploadFile } from '@/shared/hooks/useSiteConfig'
 
-interface FilePickerFieldProps {
+interface ImagePickerFieldProps {
   value: string
   onChange: (path: string) => void
-  folder?: string // Default: 'files/data'
+  folder?: string // Default: 'files/team'
   placeholder?: string
-  accept?: string // File input accept attribute
   className?: string
 }
 
-// Format file size
-function formatFileSize(bytes: number): string {
-  if (bytes < 1024) return `${bytes} B`
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`
-}
+// Image extensions
+const IMAGE_EXTENSIONS = ['.png', '.jpg', '.jpeg', '.gif', '.svg', '.webp']
 
-// Get file extension
-function getExtension(filename: string): string {
-  const parts = filename.split('.')
-  return parts.length > 1 ? parts.pop()?.toUpperCase() || '' : ''
-}
-
-export function FilePickerField({
+export function ImagePickerField({
   value,
   onChange,
-  folder = 'files/data',
-  placeholder = 'Selectionner ou uploader un fichier',
-  accept,
+  folder = 'files/team',
+  placeholder = 'Selectionner une image',
   className,
-}: FilePickerFieldProps) {
+}: ImagePickerFieldProps) {
   const [open, setOpen] = useState(false)
   const [externalUrl, setExternalUrl] = useState('')
   const fileInputRef = useRef<HTMLInputElement>(null)
@@ -57,7 +45,10 @@ export function FilePickerField({
   const { data: filesData, isLoading, refetch } = useProjectFiles(folder)
   const uploadMutation = useUploadFile()
 
-  const files = filesData?.files || []
+  // Filter only images
+  const images = (filesData?.files || []).filter((f) =>
+    IMAGE_EXTENSIONS.includes(f.extension.toLowerCase())
+  )
 
   // Handle file upload
   const handleFileUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -68,7 +59,7 @@ export function FilePickerField({
       const result = await uploadMutation.mutateAsync({ file, folder })
       await refetch()
       onChange(result.path)
-      toast.success('File uploaded', {
+      toast.success('Image uploaded', {
         description: result.filename,
       })
       setOpen(false)
@@ -89,10 +80,11 @@ export function FilePickerField({
     }
   }
 
-  // Get display name for current value
-  const displayName = value
-    ? value.split('/').pop() || value
-    : placeholder
+  // Get image URL for display
+  const getImageUrl = (path: string) => {
+    if (path.startsWith('http')) return path
+    return `/api/site/${path}`
+  }
 
   return (
     <div className={cn('flex gap-2', className)}>
@@ -100,24 +92,36 @@ export function FilePickerField({
         <PopoverTrigger asChild>
           <Button
             variant="outline"
-            className="flex-1 justify-start font-normal"
+            className="flex-1 h-auto min-h-9 justify-start font-normal p-2"
           >
             {value ? (
-              <span className="flex items-center gap-2 truncate">
-                <File className="h-4 w-4 shrink-0" />
-                <span className="truncate">{displayName}</span>
-              </span>
+              <div className="flex items-center gap-2">
+                <img
+                  src={getImageUrl(value)}
+                  alt=""
+                  className="h-8 w-8 rounded object-cover"
+                  onError={(e) => {
+                    (e.target as HTMLImageElement).style.display = 'none'
+                  }}
+                />
+                <span className="truncate text-sm">
+                  {value.split('/').pop()}
+                </span>
+              </div>
             ) : (
-              <span className="text-muted-foreground">{placeholder}</span>
+              <span className="flex items-center gap-2 text-muted-foreground">
+                <ImageIcon className="h-4 w-4" />
+                {placeholder}
+              </span>
             )}
           </Button>
         </PopoverTrigger>
-        <PopoverContent className="w-96 p-0" align="start">
-          <Tabs defaultValue="files" className="w-full">
+        <PopoverContent className="w-80 p-0" align="start">
+          <Tabs defaultValue="gallery" className="w-full">
             <TabsList className="w-full grid grid-cols-2">
-              <TabsTrigger value="files" className="gap-2">
+              <TabsTrigger value="gallery" className="gap-2">
                 <FolderOpen className="h-4 w-4" />
-                Fichiers existants
+                Galerie
               </TabsTrigger>
               <TabsTrigger value="external" className="gap-2">
                 <Link className="h-4 w-4" />
@@ -125,13 +129,13 @@ export function FilePickerField({
               </TabsTrigger>
             </TabsList>
 
-            <TabsContent value="files" className="p-2 space-y-2">
+            <TabsContent value="gallery" className="p-2 space-y-2">
               {/* Upload button */}
               <div className="flex gap-2">
                 <input
                   ref={fileInputRef}
                   type="file"
-                  accept={accept}
+                  accept="image/*"
                   onChange={handleFileUpload}
                   className="hidden"
                 />
@@ -147,23 +151,23 @@ export function FilePickerField({
                   ) : (
                     <Upload className="mr-2 h-4 w-4" />
                   )}
-                  Uploader un fichier
+                  Uploader une image
                 </Button>
               </div>
 
-              {/* Files list */}
+              {/* Images grid */}
               <ScrollArea className="h-48">
                 {isLoading ? (
                   <div className="flex items-center justify-center p-4">
                     <Loader2 className="h-5 w-5 animate-spin text-muted-foreground" />
                   </div>
-                ) : files.length === 0 ? (
+                ) : images.length === 0 ? (
                   <div className="p-4 text-center text-sm text-muted-foreground">
-                    Aucun fichier dans {folder}/
+                    Aucune image dans {folder}/
                   </div>
                 ) : (
-                  <div className="space-y-1">
-                    {files.map((file) => (
+                  <div className="grid grid-cols-4 gap-1">
+                    {images.map((file) => (
                       <button
                         key={file.path}
                         type="button"
@@ -172,20 +176,18 @@ export function FilePickerField({
                           setOpen(false)
                         }}
                         className={cn(
-                          'w-full flex items-center gap-2 rounded-md px-2 py-1.5 text-sm transition-colors hover:bg-muted',
-                          value === file.path && 'bg-primary/10 text-primary'
+                          'relative aspect-square rounded overflow-hidden border-2 transition-colors hover:border-primary/50',
+                          value === file.path
+                            ? 'border-primary'
+                            : 'border-transparent'
                         )}
+                        title={file.name}
                       >
-                        <File className="h-4 w-4 shrink-0" />
-                        <span className="flex-1 truncate text-left">{file.name}</span>
-                        <span className="text-xs text-muted-foreground shrink-0">
-                          {getExtension(file.name)}
-                        </span>
-                        {file.size && (
-                          <span className="text-xs text-muted-foreground shrink-0">
-                            {formatFileSize(file.size)}
-                          </span>
-                        )}
+                        <img
+                          src={getImageUrl(file.path)}
+                          alt={file.name}
+                          className="h-full w-full object-cover"
+                        />
                       </button>
                     ))}
                   </div>
@@ -195,13 +197,13 @@ export function FilePickerField({
 
             <TabsContent value="external" className="p-2 space-y-2">
               <p className="text-xs text-muted-foreground">
-                Entrez l'URL d'une ressource externe
+                Entrez l'URL d'une image externe
               </p>
               <div className="flex gap-2">
                 <Input
                   value={externalUrl}
                   onChange={(e) => setExternalUrl(e.target.value)}
-                  placeholder="https://example.com/data.csv"
+                  placeholder="https://example.com/photo.jpg"
                   onKeyDown={(e) => {
                     if (e.key === 'Enter') {
                       e.preventDefault()
@@ -237,4 +239,4 @@ export function FilePickerField({
   )
 }
 
-export default FilePickerField
+export default ImagePickerField
