+++ src/niamoto/gui/ui/src/features/site/components/forms/ResourcesForm.tsx
@@ -29,7 +29,7 @@ import { LocalizedInput, type LocalizedString } from '@/components/ui/localized-
 import {
   useDataContent,
   useUpdateDataContent,
-} from '@/features/site/hooks/useSiteConfig'
+} from '@/shared/hooks/useSiteConfig'
 
 // Types for resources.html context
 interface ResourceItem {
