+++ src/niamoto/gui/ui/src/features/site/views/SiteAppearancePage.tsx
@@ -1,9 +1,9 @@
-import { SitePanel } from '@/features/site/components/SitePanel'
+import { SiteBuilder } from '@/features/site/components/SiteBuilder'
 
 export default function SiteAppearancePage() {
   return (
     <div className="h-full overflow-hidden">
-      <SitePanel subSection="appearance" />
+      <SiteBuilder initialSection="appearance" />
     </div>
   )
 }
