+++ src/niamoto/gui/ui/src/features/site/views/SiteNavigationPage.tsx
@@ -1,9 +1,9 @@
-import { SitePanel } from '@/features/site/components/SitePanel'
+import { SiteBuilder } from '@/features/site/components/SiteBuilder'
 
 export default function SiteNavigationPage() {
   return (
     <div className="h-full overflow-hidden">
-      <SitePanel subSection="navigation" />
+      <SiteBuilder initialSection="navigation" />
     </div>
   )
 }
