+++ src/niamoto/gui/ui/src/features/site/views/SitePagesPage.tsx
@@ -1,12 +1,12 @@
 import { StalenessBanner } from '@/components/pipeline/StalenessBanner'
-import { SitePanel } from '@/features/site/components/SitePanel'
+import { SiteBuilder } from '@/features/site/components/SiteBuilder'
 
 export default function SitePagesPage() {
   return (
     <div className="flex h-full flex-col overflow-hidden">
       <StalenessBanner stage="site" />
       <div className="min-h-0 flex-1 overflow-hidden">
-        <SitePanel subSection="pages" />
+        <SiteBuilder initialSection="pages" />
       </div>
     </div>
   )
