--- src/niamoto/gui/ui/src/shared/i18n/index.ts
@@ -1,3 +1,0 @@
-import '@/i18n'
-
-export * from '@/i18n'
