--- src/niamoto/gui/ui/src/shared/layout/MainLayout.tsx
@@ -1,1 +1,0 @@
-export { MainLayout } from '@/components/layout/MainLayout'
