--- src/niamoto/gui/ui/src/shared/theme/index.ts
@@ -1,1 +1,0 @@
-export * from '@/components/theme'
